Fixed bug with settings menu animation.
Apparently you can't add a class via classlist.add() method to one element if it already has class so usually you'd just remove the previous class. However the settings menu needed both class on at the same time so because one was an animation, I just kept the animation final property if the other class was added. This way the settings dark/light mode was always preserved even if the animation for it was removed.